refactor(api): move sqldb.NewDatabase + migrations to dedicated db service (unblock-bne)
Extracts the canonical sqldb.NewDatabase("unblock", ...) declaration
and all nine migration files from apps/api/auth/dbhandle/ into a new
dedicated apps/api/db/ Encore service. The auth service is now an
equal database consumer, not the nominal owner of DDL for every
schema in the system.

Shape (Shape A from the bead investigation):
- apps/api/db/db.go declares //encore:service struct{} (zero APIs)
  and var DB = sqldb.NewDatabase("unblock", ...) plus init() that
  calls auth.BindDB(DB) and rbac.Bind(DB).
- apps/api/auth/db.go keeps the unblock-xuk late-bind shape
  (var db *sqldb.Database + exported BindDB hook) so plain
  go test ./apps/api/auth/... still loads without panic.
- apps/api/auth/dbhandle/ directory deleted; migrations relocated
  via git mv preserving full history.

Shape A was chosen over the investigation's recommended Shape B
("var db = sqldb.Named(\"unblock\")" mirroring org/db.go) after
empirical encore-CLI check disproved the premise that sqldb.Named
is a benign runtime lookup. sqldb.Named v1.52.1 calls doPanic at
package init exactly like sqldb.NewDatabase — Shape B would have
regressed the bead's first MUST-pass criterion (plain go test on
auth must not panic). See DECISION on unblock-bne for the full
empirical trace.

Quality gates (all clean):
- encore check: 0 errors, 30 API endpoints registered, 9 migrations
  registered from apps/api/db/migrations/.
- encore test ./...: all services green.
- go test ./apps/api/auth/...: passes without panic (xuk preserved).
- go test ./apps/api/shared/rbac/...: passes (rbac tests work).
- go test ./apps/api/org/...: fails at init (pre-existing behaviour;
  org's db.go calls sqldb.Named at package scope, not introduced or
  changed by this bead — verified via git stash).
- gofmt -l .: zero diffs.
- go vet ./...: clean.
- no_rbac_dynamic_clause + no_direct_is_ready_write analyzers: clean.
- Workspace grep: exactly ONE sqldb.NewDatabase call, in db/db.go.

Supersedes unblock-tv8.42 (the auth.go package doc-comment fix
landed here in-place; see comment on tv8.42 for closure rationale).
