+++ apps/api/auth/types/types.go
@@ -31,18 +31,24 @@
 //     annotations, or any infrastructure resource (sqldb.NewDatabase,
 //     pubsub.NewTopic, cache.NewCluster, secrets, ...). It is a pure
 //     value-type package.
-//   - SPEC §3.1 invariant: the auth service remains the sole
-//     migration-owner; sqldb.NewDatabase("unblock", ...) is still called
-//     exactly once across the workspace, in
-//     `apps/api/auth/dbhandle/dbhandle.go`. Bead unblock-xuk relocated
-//     the call from the historical apps/api/auth/db.go (where
-//     unblock-tv8.30 originally pointed) into the dbhandle leaf
-//     sub-package so plain `go test ./auth/...` no longer panics on
-//     package init — see dbhandle's package doc-comment for the full
-//     reasoning. The auth root package retains the late-bound
-//     *sqldb.Database handle (`var db *sqldb.Database` in
-//     apps/api/auth/db.go); dbhandle.init populates it via
-//     auth.BindDB.
+//   - SPEC §3.1 invariant: a dedicated apps/api/db/ service is the
+//     sole migration-owner; sqldb.NewDatabase("unblock", ...) is still
+//     called exactly once across the workspace, now in
+//     `apps/api/db/db.go`. Bead unblock-tv8.30 originally pointed at
+//     `apps/api/auth/db.go`; bead unblock-xuk relocated the call to
+//     the leaf `apps/api/auth/dbhandle/` sub-package so plain
+//     `go test ./auth/...` no longer panicked on package init; bead
+//     unblock-bne completed the decoupling by extracting the
+//     declaration AND the migration set into the dedicated db
+//     service. The auth root package retains the unblock-xuk
+//     late-bind shape (`var db *sqldb.Database` in
+//     apps/api/auth/db.go + an exported `BindDB` hook); the
+//     dedicated db service's init() calls auth.BindDB to populate
+//     it. This shape is the only one that preserves the xuk
+//     plain-`go test`-loads invariant: sqldb.Named is NOT a benign
+//     runtime lookup, it panics at package init outside the encore
+//     CLI exactly like sqldb.NewDatabase (see DECISION on bead
+//     unblock-bne for the empirical trace).
 //   - SPEC §10.1 invariant: consumer call sites continue to spell the
 //     type as `auth.Identity` (literal). They MUST NOT be re-spelled to
 //     `types.Identity` or `authtypes.Identity`. The parent `auth`
